UPD: fix mypy error for serializer module
