+++ pyproject.toml
@@ -92,7 +92,7 @@ disallow_untyped_defs = true
 disallow_incomplete_defs = false
 disallow_untyped_decorators = false
 warn_redundant_casts = true
-warn_unused_ignores = true
+warn_unused_ignores = false
 warn_return_any = false
 warn_no_return = false
 warn_unreachable = false
