+++ tests/apps/admin/article/serializers.py
@@ -1,8 +1,8 @@
-from unfazed.serializer.tortoise import TSerializer
+from unfazed.serializer import Serializer
 
 from .models import Article
 
 
-class ArticleSerializer(TSerializer):
+class ArticleSerializer(Serializer):
     class Meta:
         model = Article
