+++ tests/test_contrib/test_admin/test_registry/test_decorators.py
@@ -2,7 +2,7 @@ from tortoise import Model, fields
 
 from unfazed.contrib.admin.registry import ModelAdmin, action, admin_collector, register
 from unfazed.http import HttpRequest
-from unfazed.serializer.tortoise import TSerializer
+from unfazed.serializer import Serializer
 
 
 class Student(Model):
@@ -10,7 +10,7 @@ class Student(Model):
     age = fields.IntField()
 
 
-class StudentSerializer(TSerializer):
+class StudenSerializer(Serializer):
     class Meta:
         model = Student
 
@@ -18,7 +18,7 @@ class StudentSerializer(TSerializer):
 async def test_decorator() -> None:
     admin_collector.clear()
 
-    @register(StudentSerializer)
+    @register(StudenSerializer)
     class StudentAdmin(ModelAdmin):
         @action(name="test_action", confirm=True)
         def test_action(self, request: HttpRequest | None = None) -> str:
@@ -31,7 +31,7 @@ async def test_decorator() -> None:
     assert "StudentAdmin" in admin_collector
 
     ins: ModelAdmin = admin_collector["StudentAdmin"]
-    assert ins.serializer == StudentSerializer
+    assert ins.serializer == StudenSerializer
 
     actions = ins.get_actions()
