+++ tests/test_serializer/test_serializer_fields.py
@@ -8,7 +8,7 @@ from tortoise import Model, fields
 
 from tests.apps.serializer.models import Bag, Course, Student
 from tests.apps.serializer.models import StudentProfile as Profile
-from unfazed.serializer.tortoise import TSerializer, create_common_field
+from unfazed.serializer import Serializer, create_common_field
 
 
 class Sex(StrEnum):
@@ -41,13 +41,13 @@ class User(Model):
     uuid = fields.UUIDField()
 
 
-class UserSerializer(TSerializer):
+class UserSerializer(Serializer):
     class Meta:
         model = User
         exclude = ["is_active"]
 
 
-class UserSerializer2(TSerializer):
+class UserSerializer2(Serializer):
     # override field
     height: int = Field(...)
 
@@ -58,7 +58,7 @@ class UserSerializer2(TSerializer):
         model = User
 
 
-class UserSerializer3(TSerializer):
+class UserSerializer3(Serializer):
     class Meta:
         model = User
         include = ["height"]
@@ -222,26 +222,26 @@ def test_create_common_fields() -> None:
 
 
 def test_create_relational_fields() -> None:
-    class StudentSerializer(TSerializer):
+    class StudenSerializer(Serializer):
         class Meta:
             model = Student
 
     # db fields
-    assert "id" in StudentSerializer.model_fields
-    assert "name" in StudentSerializer.model_fields
-    assert "age" in StudentSerializer.model_fields
+    assert "id" in StudenSerializer.model_fields
+    assert "name" in StudenSerializer.model_fields
+    assert "age" in StudenSerializer.model_fields
 
     # m2m fields
-    assert "courses" in StudentSerializer.model_fields
+    assert "courses" in StudenSerializer.model_fields
 
     # fk fields
-    assert "bags" in StudentSerializer.model_fields
+    assert "bags" in StudenSerializer.model_fields
     # one to one fieldss
-    assert "profile" in StudentSerializer.model_fields
+    assert "profile" in StudenSerializer.model_fields
 
     # init model
 
-    student = StudentSerializer(
+    student = StudenSerializer(
         name="student1",
         age=18,
         courses=[
@@ -261,7 +261,7 @@ def test_create_relational_fields() -> None:
     assert student.profile.nickname == "nick1"
 
     # relation fields default to None
-    student = StudentSerializer(name="student1", age=18)
+    student = StudenSerializer(name="student1", age=18)
 
     assert student.name == "student1"
     assert student.courses is None
@@ -269,7 +269,7 @@ def test_create_relational_fields() -> None:
     assert student.profile is None
 
     # test profile serializer
-    class ProfileSerializer(TSerializer):
+    class ProfileSerializer(Serializer):
         class Meta:
             model = Profile
 
@@ -289,7 +289,7 @@ def test_create_relational_fields() -> None:
 
     # test bag serializer
 
-    class BagSerializer(TSerializer):
+    class BagSerializer(Serializer):
         class Meta:
             model = Bag
 
@@ -306,7 +306,7 @@ def test_create_relational_fields() -> None:
     assert bag.student_id == 1
 
     # test course serializer
-    class CourseSerializer(TSerializer):
+    class CourseSerializer(Serializer):
         class Meta:
             model = Course
 
@@ -331,20 +331,20 @@ def test_create_relational_fields() -> None:
 def test_meta() -> None:
     with pytest.raises(ValueError):
 
-        class NoModelSerializer(TSerializer):
+        class NoModelSerializer(Serializer):
             class Meta:
                 foo = "bar"
 
     # test include exclude conflict
     with pytest.raises(ValueError):
 
-        class IncludeExcludeConfict(TSerializer):
+        class IncludeExcludeConfict(Serializer):
             class Meta:
                 model = Student
                 include = ["name"]
                 exclude = ["name"]
 
-    class IncludeSerializer(TSerializer):
+    class IncludeSerializer(Serializer):
         class Meta:
             model = Student
             include = ["name"]
@@ -354,5 +354,5 @@ def test_meta() -> None:
 
     with pytest.raises(ValueError):
 
-        class NoMetaSerializer(TSerializer):
+        class NoMetaSerializer(Serializer):
             pass
