+++ unfazed/serializer/__init__.py
@@ -1,3 +1,4 @@
-from .base import BaseSerializer
+from .base import Serializer
+from .utils import create_common_field
 
-__all__ = ["BaseSerializer"]
+__all__ = ["Serializer", "create_common_field"]
