--- unfazed/serializer/tortoise/__init__.py
@@ -1,3 +1,0 @@
-from .serializer import TSerializer, create_common_field
-
-__all__ = ["TSerializer", "create_common_field"]
